+++ src/swganh/messages/update_transform_with_parent_message.h
@@ -45,9 +45,9 @@ struct UpdateTransformWithParentMessage : public swganh::messages::BaseSwgMessag
     void onSerialize(anh::ByteBuffer& buffer) const {
         buffer.write(cell_id);
         buffer.write(object_id);
-        buffer.write<int16_t>(position.x * 8.0f + 0.5f);
-        buffer.write<int16_t>(position.y * 8.0f + 0.5f);
-        buffer.write<int16_t>(position.z * 8.0f + 0.5f);
+        buffer.write<int16_t>(static_cast<int16_t>(position.x * 8.0f + 0.5f));
+        buffer.write<int16_t>(static_cast<int16_t>(position.y * 8.0f + 0.5f));
+        buffer.write<int16_t>(static_cast<int16_t>(position.z * 8.0f + 0.5f));
         buffer.write(update_counter);
         buffer.write(static_cast<uint8_t>(0));
         buffer.write(heading);
